requirements-with-cupy.txt
torch
numpy
einops
opencv-python
kornia
scipy
Pillow
torchvision
tqdm
cupy-cuda13x
